golibmc_test: Fix fragile test case "testQuit"
The bug is not stable to reproduce on Linux and
it CANNOT be reproduced on OS X. A possible explanation to this
phenomenon is:

Sending version command will not increase the `curr_connections` metric,
and only data I/O commands like Set/Get/Delete commands will do. The
idea of this commit is to make sure the current connection to
target memcached server is established and confirmed by the server-end
(curr_connections + 1).